# packages/api-client/src/common/models/RegisterRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegisterRequest = {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
};