# packages/api-client/src/common/models/ResetPasswordRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ResetPasswordRequest = {
  token: string;
  password: string;
};