# packages/api-client/src/common/models/UpdateUserRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateUserRequest = {
  firstName?: string;
  lastName?: string;
  role?: 'admin' | 'user';
};